Add dynamic capacity pools
This commit adds the "dynamic capacity pools" feature. Unlike in the old
management, cap pools are calculated on demand from free space/inodes
information. This results in a somewhat complex query (which is also
executed multiple times). This is probably ok, but could also be cached
if actually causing a db bottleneck.

Also:
* Refactor get_storage_pools response building
* Fix some small bugs

Missing: Tests that is actually works as intended (smoke test has been
done)