add quota config options
It is necessary to be able to specify IDs to be checked using the old
methods: via file or via range or using system users. All of these are
usable now, the old manual mechanism has been removed.

This is not ideal: The plan is to not having to specify
any ID at all, instead just query everything that is available on the
storage targets and has quota information.