Allow registration of multiple Nodes via RegisterNodeMsg providing same alias
Nodes use their hostname as default for the alias / node string id when registering with RegisterNodeMsg / Heartbeat.
Since aliases are globally unique, registering multiple servers running on the same host didn't work.

This patch extends the aliases with nodetype and ID if it already exists in the database - no more registration clash.

Also, the alias is not updated anymore if the node already exists (we want the user to be able to manually rename them)