Fix alias behavior
Prepend node aliases / nodeStrId as they come in on RegisterNode with
nodetype and ID (instead of appending only if there are duplicated).
This makes sure that the alias always  starts with a letter as it is
required. Add "target_" in front of storage targets on registration.
This makes sure these also start with a letter.