Implement v7 import functionality
Management can now import v7 management data into a new database.
Certain ephemeral data is omitted, it will be filled automatically
during operation.

The import is only supported from the latest v7 management format (not
sure when this was last changed, but it should be fine for 7.2 and
later).

Upgrading requires a clean shutdown of the old BeeGFS, with all targets
being in GOOD state and all clients being unmounted.