Replace From/TryFrom from impl_enum_to_int! macro and rename
These were meant for converting a BeeGFS enum into an integer
for BeeMsg. That turns out to be dangerous though because one can
accidentally use this conversion for proto messages or other things
which might not necessarily use the same integer representations.

To avoid bugs we replace the TryFrom impl with a custom trait that
provides explicit from_/into_ bee_serde functions.