Replace "enum" TEXT fields with INTEGER and types tables
Using "free" values as "enums" is considered bad style as they can
easily be set to an invalid value. Checks could prevent that, but
are a pain to deal with if they should ever change.

Instead add new _types tables that basically represent the enums and
use foreign keys to reference them. This makes sure that "enum" entries
are always valid.

The _types tables should, of course, never be touched manually unless
new enum variants are added or old ones are removed.