Add SAFETY comments, add lint to deny unsafe without comment
It's pretty important that a caller of an unsafe function ensures that
it's contract is satisfied. To make this more clear to the developer as
well as others, we now require a `SAFETY: ` comment on all `unsafe` blocks.
Not to be confused with a `# Safety` section on unsafe functions or
structs which _defines_ the contract the caller has to fulfill, a `SAFETY:`
comment should be used to reason about why calling this here is actually
safe. Or why it is not but we accept the risk anyway.

Also

* Make `LoadedLibrary::new()` unsafe as calling this actually is - it
  being safe totally depends on the actual library behind `path` being
  the correct one and satisfying `LoadedLibraries` contract. There is
  no way the wrapper could enforce this internally, neither as compile
  nor at runtime. Therefore, is has to marked as unsafe and require the
  caller to provide the correct library.
* Add `SAFETY:` comments to all existing `unsafe` blocks
* Add `# Safety` comments to some unsafe licensing fns and structs
* Some small code improvements in `license.rs` (removing unnecessary
  maps and matches)