Use a macro to generate the grpc handler impls
All of the implemented functions in ManagementService basically do
the same, with small differences for unary and streaming rpcs - they
forward the request to the actual handler function and process any
error returned. Using a macro makes it easier to get an overview what
impls are there.

The macro requires specifying the message types and the
function where the actual handling happens. The generated code handles
any returned error and now also provides support for responding with
specified status codes. If the error coming out of the given handler
function contains a `tonic::Status`, that one is directly returned
to the requestor. If not, a generic `Status::Internal` containing the
stringified error is sent back. This allows usage of `?` but also gives
control over the status code sent back if needed.