Check no clients being connected before mirroring the root inode
This is important to prevent data corruption but has been forgotten.
Note that since mirroring the root inode is a two step process, there is
still a small chance for a client being mounted again before the action
is taken on the root meta server. In the end, it's the administrators
responsibility to not let any client mount during that process.