Logging improvements
* Disable logging the target to stdout. These are not very useful atm as
they print the module the macro was invoked by default. But this happens
outside of the handlers for most errors and is therefore misleading and
only blows up the log output.
* If a gRPC response stream aborts early due to the receiver closing the
channel, do not log it as error as this usually happens when a client
cancels receiving the stream early (which is usually not a problem).
* Small gRPC log output improvements