Remove switchover on primary being `needs resync`
As discussed, this is somehow dangerous and a switchover is only safe
if all nodes are "guaranteed" (as far as we can guarantee it) to know
the primary as "probably offline".

Since nobody really knows the condition oh which this can happen, better
be safe than sorry and remove it. If things don't work as expected, we
can re-add it.