Overhaul db initialization and v7 import
Now the database will be created in memory first and only be written to
disk after the whole process has been succeeded. This prevents
accidental use of a fresh database with an existing BeeGFS system if the
v7 import fails and also a potential partly initialized db if one of the
init steps fails.

--import-from-v7 now also works by itself, without having to --init
first.