Add options to daemonize the process
Includes a pid file.

This is only implemented because apparently needed in certain
circumstances. Startup logging and result is unavailable in this
mode, e.g. there is no feedback if the management was actually started
successfully. One has to investigate the log journal to check and/or
check if the process is running.

Note that it still requires systemd / the journal for logging even if it
doesn't need it for running. I can confirm that logging to the journal
still works even if there is no beegfs-mgmtd unit installed (management
uses a local socket to send the log messages to the journal). Thus, the
options are hidden from the help output. Normal users should use systemd
instead.

The options are: --daemonize for daemonizing and --daemonize-pid-file to
set the pid file location.