Add rust-toolchain.toml and adapt CI
This fixes the build toolchain to a specific Rust version, so it doesn't
automatically use the newest one (excluding patches) anymore. Although
Rust releases are backward compatible, there might be new lints that
issue new warnings which might the CI fail after a new release. By
fixing the toolchain to use, this shouldn't happen anymore.

Also, remove `--locked` from `cargo generate-rpm` as it is not a valid
flag there.