Add a debian style copyright file to the rpm package
This is a manually created file that matches the files from the core
repo. The deb pacakge one is auto generated and looks a bit different.

In addition, change the target paths of the thirdparty-licenses.html to
match the copyright file path and also the convention of the other
packages.