Store and report reachability per target
Management regularly sends a `ChangeTargetConsistencyStateMsg`, which
contains one entry per known target. Use this to have separate last
contact times per target. Query this time instead of the nodes time when
calculating reachability states. This way, a missing/disabled target
will not be considered online.

Note that this mechanism is far from being perfect, but the best we can
get with the current protocol.