Add Zlib license to allowed licenses (ThinkParQ/beegfs-rs#173)
* Add Zlib license to allowed licenses

The recent update to the new Rust editions and new package dependencies
added an indirect dependency on "foldhash" via "hashbrown". The
"foldhash" crate is distributed under the Zlib license, which we didn't
have in our whitelist yet.

The Zlib license itself (http://www.zlib.net/zlib_license.html) seems to
be a more permissive version of the BSD-3-Clause license, which we also
allow. It explicitly permits use of the code in commercial products, so
I think we are good allowing it.

* make sure license checks run before merge