Revert edition to 2021 and fix CI package builds (ThinkParQ/beegfs-rs#174)
This fixes a weird issue with the makefile where the TARGET_FLAG variable
wasn't properly set which caused build failures.

Something with the new .SILENT target at the top must be messing with
the scoping and changes the scope of that variable to a local scope,
causing it to be empty in the later build steps.

I haven't fully understood what exactly is going wrong, but in the
interest of time, this "fix" will have to do for now.