fix: do not abort on deserialization allocation failure
The standard library aborts (not panics) when a memory allocation fails.

If the deserializer retrieves corrupted data, it might read out some big
value for a sequences or maps length field which causes the reservation
done with `with_capacity()` to abort the process.

To avoid this, we use `new()` and `try_reserve` instead as this should
just return an error if the allocator fails.