fix: skip setting the root inode on the first meta target when importing v7 data
This messes with buddy group import later. Instead leave it undefined,
the import process sets it manually at the end.

Note that this fix is a bit dirty by exposing a previously private
method. But since the plan is to get rid of the "database layer" anyway
during later refactoring, it's fine.