fix: add option to start a db write transaction without fsyncing
This avoids fsync to disk after the transaction and thus greatly reduces
disk load on some heavy used handlers. While a database in WAL mode is
guaranteed to be safe from corruption with this setting, it's not that
durable - the transaction could be lost on a crash or power outage.